Migrates pull.net package into the new plugin layout
