+++ tasks/linting.py
@@ -37,12 +37,10 @@ def mypy(ctx):
         "{}/engines/*.py".format(SOURCE_PATH),
         "{}/plugins/__init__.py".format(SOURCE_PATH),
         "{}/plugins/pull/__init__.py".format(SOURCE_PATH),
-        "{}/plugins/pull/monitor/stats.py".format(SOURCE_PATH),
-        "{}/plugins/pull/simple/count.py".format(SOURCE_PATH),
-        "{}/plugins/pull/simple/cron.py".format(SOURCE_PATH),
-        "{}/plugins/pull/simple/repeat.py".format(SOURCE_PATH),
-        "{}/plugins/pull/simple/run_once.py".format(SOURCE_PATH),
-        "{}/plugins/pull/tracking/fritzboxtracker.py".format(SOURCE_PATH),
+        "{}/plugins/pull/monitor/*.py".format(SOURCE_PATH),
+        "{}/plugins/pull/net/*.py".format(SOURCE_PATH),
+        "{}/plugins/pull/simple/*.py".format(SOURCE_PATH),
+        "{}/plugins/pull/tracking/*.py".format(SOURCE_PATH),
         "{}/plugins/push/__init__.py".format(SOURCE_PATH),
         "{}/plugins/udf/__init__.py".format(SOURCE_PATH),
         "{}/shared/*.py".format(SOURCE_PATH),
